backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-Login
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
